requirements.txt
accelerate
beartype
einops
ema-pytorch
encodec
fairseq
joblib
lion-pytorch
local-attention
scikit-learn
sentencepiece
torchaudio
transformers
tqdm
vector-quantize-pytorch
librosa
tensorboardX
https://download.pytorch.org/whl/cu111/torch-1.10.2%2Bcu111-cp39-cp39-linux_x86_64.whl
https://download.pytorch.org/whl/cu111/torchaudio-0.10.2%2Bcu111-cp39-cp39-linux_x86_64.whl
